Rename src.server.{load_balancing => block_selection}
