+++ src/server/block_selection.py
